backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# AI/ML dependencies for multi-agent system
# unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git  # Commented out due to Python 3.13 compatibility
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
peft>=0.7.0
huggingface_hub>=0.19.0 